requirements.txt
flask
streamlit
pandas~=1.5.3
numpy~=1.24.1
pyyaml==6.0
tensorflow~=2.11.0
requests==2.28.1
pydantic==1.10.2
uvicorn==0.15.0
pillow==9.2.0
